Cast to float

+++ mnist_dcgan/dcgan.py
@@ -178,9 +178,9 @@ if __name__ == '__main__':
             netD.zero_grad()
             real_cpu = data[0].to(device)
             batch_size = real_cpu.size(0)
-            label = torch.full((batch_size,), real_label, device=device)
+            label = torch.full((batch_size,), real_label, device=device).float()
 
-            output = netD(real_cpu)
+            output = netD(real_cpu).float()
             errD_real = criterion(output, label)
             errD_real.backward()
             D_x = output.mean().item()
